fix: remove tab index off Modal
Authored by @dgiraldo313. Merged by @BeniCheni.

This fixes the issue where tabIndex is stuck on the close button on the modal and doesn't allow for other tab-able items to be focused. This issue is caused by a tabIndex="1", which prevents other actionable items to be focused, since they all have a tabIndex="0" by default.

PR-URL: #537
Reviewed-By: @sdalonzo, @BeniCheni